add validationRules.ts in utils(centralized backend validationErrors),update user friendly error message in SignUpPage.tsx and update test SignUpPage.spec.tsx
